day24 part1

+++ 2023/day24/solver_test.go
@@ -8,31 +8,31 @@ func TestSolver1(t *testing.T) {
 
 	puzzle := LoadPuzzle("test.txt")
 
-	test := func(puzzle []Hail, area []uint64, expected int) {
+	test := func(puzzle []Hail, area []float64, expected int) {
 		result := Solver1(puzzle, area)
 		if result != expected {
 			t.Errorf("Expected %d, result %d", expected, result)
 		}
 	}
 
-	test([]Hail{}, []uint64{}, 0)
-	test(puzzle[0:1], []uint64{7, 27}, 0)
-	test(puzzle[0:2], []uint64{7, 27}, 1)
-	test(puzzle[0:3], []uint64{7, 27}, 2)
-	test(puzzle, []uint64{7, 27}, 2)
+	test([]Hail{}, []float64{}, 0)
+	test(puzzle[0:1], []float64{7, 27}, 0)
+	test(puzzle[0:2], []float64{7, 27}, 1)
+	test(puzzle[0:3], []float64{7, 27}, 2)
+	test(puzzle, []float64{7, 27}, 2)
 }
 
 func TestSolver2(t *testing.T) {
 
 	puzzle := LoadPuzzle("test.txt")
 
-	test := func(puzzle []Hail, area []uint64, expected int) {
+	test := func(puzzle []Hail, area []float64, expected int) {
 		result := Solver2(puzzle)
 		if result != expected {
 			t.Errorf("Expected %d, result %d", expected, result)
 		}
 	}
 
-	test([]Hail{}, []uint64{}, 0)
-	test(puzzle, []uint64{7, 27}, 0)
+	test([]Hail{}, []float64{}, 0)
+	test(puzzle, []float64{7, 27}, 0)
 }
